refactor, add doc strings, clean up
